delete pages/results/index
